+++ examples/map_key_serialization.rs
@@ -6,7 +6,7 @@
 use serde_derive::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::error::Error;
-use winreg::enums::*;
+use winreg::HKCU;
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
 struct Coords {
@@ -27,8 +27,7 @@ struct Rectangle {
 }
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let hkcu = winreg::RegKey::predef(HKEY_CURRENT_USER);
-    let (key, _disp) = hkcu.create_subkey("Software\\RustEncodeMapKey")?;
+    let (key, _disp) = HKCU.create_subkey("Software\\RustEncodeMapKey")?;
     let mut v1 = HashMap::new();
     v1.insert(
         "first".to_owned(),