+++ examples/serialization.rs
@@ -6,7 +6,7 @@
 use serde_derive::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::error::Error;
-use winreg::enums::*;
+use winreg::HKCU;
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
 struct Coords {
@@ -52,8 +52,7 @@ struct Test {
 }
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let hkcu = winreg::RegKey::predef(HKEY_CURRENT_USER);
-    let (key, _disp) = hkcu.create_subkey("Software\\RustEncode")?;
+    let (key, _disp) = HKCU.create_subkey("Software\\RustEncode")?;
 
     let mut map = HashMap::new();
     map.insert("".to_owned(), 0); // empty name becomes the (Default) value in the registry
@@ -86,7 +85,13 @@ fn main() -> Result<(), Box<dyn Error>> {
         t_f32: 3.15,
     };
 
+    // to just write the data from structure without touching other
+    // subkeys/values in the target key:
     key.encode(&v1)?;
+
+    // or
+    // to wipe out everything under the target key:
+    // key.encode_destructive(&v1)?;
 
     let v2: Test = key.decode()?;
     println!("Decoded {:?}", v2);